Moving controller concerns to lib
