--- app/controllers/concerns/with_exams_validations.rb
@@ -1,6 +1,0 @@
-module WithExamsValidations
-  def validate_accessible(item)
-    item.access!(current_user) unless (current_user? && current_user.teacher_here?)
-  end
-
-end
