--- app/controllers/concerns/with_organization_chooser.rb
@@ -1,7 +1,0 @@
-module WithOrganizationChooser
-
-  def should_choose_organization?
-    current_user? && current_user.has_accessible_organizations? && Mumukit::Platform.implicit_organization?(request)
-  end
-
-end
