+++ app/controllers/exercises_controller.rb
@@ -1,6 +1,4 @@
 class ExercisesController < ApplicationController
-  include WithExamsValidations
-
   before_action :set_guide, only: :show
   before_action :set_default_content, only: :show, if: :current_user?
   before_action :set_assignment, only: :show, if: :current_user?
