+++ lib/mumuki/laboratory.rb
@@ -22,3 +22,4 @@ require_relative './laboratory/extensions'
 require_relative './laboratory/exceptions'
 require_relative './laboratory/status'
 require_relative './laboratory/evaluation'
+require_relative './laboratory/controllers'
