+++ lib/mumuki/laboratory/controllers.rb
@@ -1,0 +1,9 @@
+module Mumuki::Laboratory::Controllers
+end
+
+require_relative './controllers/authorization'
+require_relative './controllers/current_organization'
+require_relative './controllers/dynamic_errors'
+require_relative './controllers/messages'
+require_relative './controllers/nested_in_exercise'
+require_relative './controllers/results_rendering'
