+++ lib/mumuki/laboratory/controllers/current_organization.rb
@@ -1,4 +1,4 @@
-module WithOrganization
+module Mumuki::Laboratory::Controllers::CurrentOrganization
   def set_current_organization!
     Organization.find_by!(name: organization_name).switch!
   rescue => e
